Refactor: Modernize project structure and add testing
This commit refactors the main Flask application to improve modularity and maintainability, aligning it with modern software standards.

The key changes include:
- Separating the core application logic from the main `flask_app.py` file into dedicated service modules:
  - `app/services/spotify_service.py`: Handles Spotify authentication and API client creation.
  - `app/services/analysis_service.py`: Contains all music analysis logic, including Spotify data processing, librosa-based audio analysis, and optional LLM integration.
- Cleaning up the `requirements.txt` file by removing a duplicate `openai` dependency.
- Introducing a basic testing structure with a new `tests` directory and a sample test file (`tests/test_analysis_service.py`) to encourage future test-driven development.

These changes make the codebase easier to understand, test, and extend, which were the primary goals of the modernization effort.
