+++ requirements.txt
@@ -8,7 +8,5 @@ langdetect
 numpy<2
 scipy>=1.13.1
 librosa==0.10.2.post1
-openai>=1.3.0
-
 openai>=1.40.0
